refactor: HashMap->Map 변경
